+++ rosetta_source/src/apps/pilot/will/symdock_enum_3_1.cc
@@ -166,8 +166,8 @@ sicfast(
   ymx = min(ymx,ymx1); ymn = max(ymn,ymn1);
 
 
-  int xlb = floor(xmn/BIN)-2; int xub = ceil(xmx/BIN)+2; // one extra on each side for correctness,
-  int ylb = floor(ymn/BIN)-2; int yub = ceil(ymx/BIN)+2; // and one extra for outside atoms
+  int xlb = (int)floor(xmn/BIN)-2; int xub = (int)ceil(xmx/BIN)+2; // one extra on each side for correctness,
+  int ylb = (int)floor(ymn/BIN)-2; int yub = (int)ceil(ymx/BIN)+2; // and one extra for outside atoms
 
   // TR << "BOUNDS " << xmn << " " << xmx << " " << ymn << " " << ymx << std::endl;
   // TR << "BOUNDS " << xlb << " " << xub << " " << ylb << " " << yub << std::endl;
@@ -580,8 +580,8 @@ void run( Size itrifile, Size ipntfile ) {
           // for(iori = 0; iori < 360; iori+=ANGLE_INCR)
           while(ori_stage < 5) {
             if(newstage) {
-              if( ori_stage == 1 || ori_stage == 2 ) iori = ( 90.0+double(ANGLE_INCR)/2.0+angle_degrees(taxs,Vecf(0,0,0),paxs));
-              if( ori_stage == 3 || ori_stage == 4 ) iori = (270.0+double(ANGLE_INCR)/2.0+angle_degrees(taxs,Vecf(0,0,0),paxs));
+              if( ori_stage == 1 || ori_stage == 2 ) iori = (int)( 90.0+double(ANGLE_INCR)/2.0+angle_degrees(taxs,Vecf(0,0,0),paxs));
+              if( ori_stage == 3 || ori_stage == 4 ) iori = (int)(270.0+double(ANGLE_INCR)/2.0+angle_degrees(taxs,Vecf(0,0,0),paxs));
               iori = (iori / ANGLE_INCR) * ANGLE_INCR; // round to closest multiple of angle incr
               if( ori_stage == 2 || ori_stage == 4 ) iori -= ANGLE_INCR;
               newstage = false;
@@ -605,8 +605,8 @@ void run( Size itrifile, Size ipntfile ) {
             if( w > 0 ) {
               pntmn = pntmnpos[ipnt/ANGLE_INCR+1];
               trimn = trimnpos[itri/ANGLE_INCR+1];
-              int dp = (dpnt-pntmn)*10+1;
-              int dt = (dtri-trimn)*10+1;
+              int dp = (int)(dpnt-pntmn)*10+1;
+              int dt = (int)(dtri-trimn)*10+1;
               if( dp < 1 ) { ori_stage++; newstage=true; continue; };
               if( dt < 1 ) { ori_stage++; newstage=true; continue; };
               // if(ipnt==18 && itri==72 && iori==276)
@@ -618,8 +618,8 @@ void run( Size itrifile, Size ipntfile ) {
             } else {
               pntmn = pntmnneg[ipnt/ANGLE_INCR+1];
               trimn = trimnneg[itri/ANGLE_INCR+1];
-              int dp = (-dpnt+pntmn)*10+1;
-              int dt = (-dtri+trimn)*10+1;
+              int dp = (int)(-dpnt+pntmn)*10+1;
+              int dt = (int)(-dtri+trimn)*10+1;
               if( dp < 1 ) { ori_stage++; newstage=true; continue; };
               if( dt < 1 ) { ori_stage++; newstage=true; continue; };
               // if(ipnt==18 && itri==72 && iori==276)
@@ -911,8 +911,8 @@ void run( Size itrifile, Size ipntfile ) {
               if( w > 0 ) {
                 pntmn = pntmnpos[ipnt/ANGLE_INCR+1];
                 trimn = trimnpos[itri/ANGLE_INCR+1];
-                int dp = (dpnt-pntmn)*10+1;
-                int dt = (dtri-trimn)*10+1;
+                int dp = (int)(dpnt-pntmn)*10+1;
+                int dt = (int)(dtri-trimn)*10+1;
                 if( dp < 1 ) { continue; };
                 if( dt < 1 ) { continue; };
                 // if(ipnt==18 && itri==72 && iori==276)
@@ -1005,8 +1005,8 @@ void run( Size itrifile, Size ipntfile ) {
         if( w > 0 ) {
           pntmn = pntmnpos[ipnt/ANGLE_INCR+1];
           trimn = trimnpos[itri/ANGLE_INCR+1];
-          int dp = (dpnt-pntmn)*10+1;
-          int dt = (dtri-trimn)*10+1;
+          int dp = (int)(dpnt-pntmn)*10+1;
+          int dt = (int)(dtri-trimn)*10+1;
           if( dp < 1 ) { continue; };
           if( dt < 1 ) { continue; };
           // if(ipnt==18 && itri==72 && iori==276)
@@ -1018,8 +1018,8 @@ void run( Size itrifile, Size ipntfile ) {
         } else {
           pntmn = pntmnneg[ipnt/ANGLE_INCR+1];
           trimn = trimnneg[itri/ANGLE_INCR+1];
-          int dp = (-dpnt+pntmn)*10+1;
-          int dt = (-dtri+trimn)*10+1;
+          int dp = (int)(-dpnt+pntmn)*10+1;
+          int dt = (int)(-dtri+trimn)*10+1;
           if( dp < 1 ) { continue; };
           if( dt < 1 ) { continue; };
           // if(ipnt==18 && itri==72 && iori==276)
